+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog 
 ## Version 0.62.8
 - fixed stations history displaying "null" when warnings are disabled and municipality names instead of the station name are displayed
+- remove pre-fetching of map images to speed up main ui on older devices
+- added information about uv index
 
 ## Version 0.62.7
 - fixed maps activity crashing when no internet connection and/or no pollen data present